pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-math-assistant"
version = "0.1.0"
description = "AI Math Assistant with FastAPI backend and React frontend."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "langchain",
    "langchain-openai",
    "langchain-community",
    "sympy",
    "python-dotenv",
    "openai",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/yourname/ai-math-assistant"
"Source" = "https://github.com/yourname/ai-math-assistant"

[project.optional-dependencies]
dev = ["pytest", "requests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_math_assistant*"]
